requirements.txt
google-cloud==0.34.0
google-cloud-bigquery==0.28.0
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.0.3
gunicorn==20.1.0
